refactor: 풀이 기록 조회 기능을 statistic 도메인으로 이동
SolvingResultService, MySolveRecordDto, QuestionSolvingRecordController,
UserSolvingResultApiResponse 및 관련 테스트를 solve 도메인에서
신설한 statistic 도메인으로 옮긴다.
