+++ src/main/java/com/coniv/mait/domain/statistic/service/SolvingResultService.java
@@ -1,4 +1,4 @@
-package com.coniv.mait.domain.solve.service;
+package com.coniv.mait.domain.statistic.service;
 
 import java.util.List;
 import java.util.Map;
@@ -14,8 +14,8 @@ import com.coniv.mait.domain.solve.entity.AnswerSubmitRecordEntity;
 import com.coniv.mait.domain.solve.exception.QuestionSolveExceptionCode;
 import com.coniv.mait.domain.solve.exception.QuestionSolvingException;
 import com.coniv.mait.domain.solve.service.component.AnswerSubmitRecordReader;
-import com.coniv.mait.domain.solve.service.dto.MySolveRecordDto;
 import com.coniv.mait.domain.solve.service.dto.QuestionSolveResultDto;
+import com.coniv.mait.domain.statistic.service.dto.MySolveRecordDto;
 import com.coniv.mait.domain.user.service.component.TeamRoleValidator;
 import com.coniv.mait.global.auth.model.MaitUser;
